feat: move to float

+++ midealocal/devices/ac/__init__.py
@@ -366,7 +366,7 @@ class MideaACDevice(MideaDevice):
 
     def set_target_temperature(
         self,
-        target_temperature: int,
+        target_temperature: float,
         mode: int | None,
         zone: int | None = None,  # noqa: ARG002
     ) -> None:
